Fine-tune and document ordering of Prefix/MaxLenPrefix
Instead of relying on the derived ordering and the FamilyAndLen u8 on
its own, this implements Ord and PartialOrd for both Prefix and
MaxLenPrefix explicitly. This way we can include the actual Bits in the
comparison, which is necessary to determine whether one prefix is a
more-specific of the other and order them accordingly.
Additionally, this takes into account the (presence of) max length in
MaxLenPrefix.