Introduce Nlri::Unicast/Multicast to replace Nlri::Basic
The `Nlri::Basic` enum variant was used to carry both unicast and
multicast SAFI NLRI. Replacing the `Basic` variant with explicit
`Unicast` and `Multicast` variants is less error-prone for the user.
Additionally, this commit removes methods from `Nlri` that hid similar
details, e.g. the `prefix()` allowed the user to extract the prefix
without knowing the exact variant of `Nlri`, so a user might unknowingly
be looking at and using a prefix coming from a MplsVpn NLRI for example
while they are actually only interested in say, unicast announcements.